backend/requirements.txt
fastapi
uvicorn
python-multipart
httpx
python-dotenv
groq